+++ smart-campus-service-api/src/main/java/hu/unideb/smartcampus/service/api/domain/response/wrapper/inner/InstructorWrapper.java
@@ -1,5 +1,6 @@
-package hu.unideb.smartcampus.service.api.domain.response.wrapper;
+package hu.unideb.smartcampus.service.api.domain.response.wrapper.inner;
 
+import hu.unideb.smartcampus.service.api.domain.response.wrapper.BaseWrapper;
 import lombok.Builder;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
