+++ smart-campus-service-api/src/main/java/hu/unideb/smartcampus/service/api/domain/response/wrapper/inner/SubjectWrapper.java
@@ -1,7 +1,8 @@
-package hu.unideb.smartcampus.service.api.domain.response.wrapper;
+package hu.unideb.smartcampus.service.api.domain.response.wrapper.inner;
 
 import java.util.List;
 
+import hu.unideb.smartcampus.service.api.domain.response.wrapper.BaseWrapper;
 import lombok.Builder;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
