+++ smart-campus-service-api/src/main/java/hu/unideb/smartcampus/service/api/domain/response/wrapper/inner/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * These wrapper classes are members of another wrapper classes, so they do not have to set the
+ * BaseWrapper's class messageType.
+ */
+package hu.unideb.smartcampus.service.api.domain.response.wrapper.inner;
