+++ smart-campus-service-api/src/main/java/hu/unideb/smartcampus/service/api/domain/response/wrapper/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * These wrapper classes will be responsible to provide response to users over XMPP protocoll, these
+ * wrappers would transfer the data/messages, so they HAVE TO set the BaseWrapper's messageType.
+ */
+package hu.unideb.smartcampus.service.api.domain.response.wrapper;
