+++ smart-campus-service-impl/src/main/java/hu/unideb/smartcampus/service/api/requestprocess/RetrieveSubjectsRequestServiceImpl.java
@@ -18,9 +18,9 @@ import hu.unideb.smartcampus.persistence.entity.SubjectEntity;
 import hu.unideb.smartcampus.persistence.repository.InstructorRepository;
 import hu.unideb.smartcampus.persistence.repository.UserRepository;
 import hu.unideb.smartcampus.service.api.MessageProcessingClass;
-import hu.unideb.smartcampus.service.api.domain.response.wrapper.InstructorWrapper;
 import hu.unideb.smartcampus.service.api.domain.response.wrapper.SubjectRetrievalResponseWrapper;
-import hu.unideb.smartcampus.service.api.domain.response.wrapper.SubjectWrapper;
+import hu.unideb.smartcampus.service.api.domain.response.wrapper.inner.InstructorWrapper;
+import hu.unideb.smartcampus.service.api.domain.response.wrapper.inner.SubjectWrapper;
 import hu.unideb.smartcampus.shared.requestmessages.BaseRequestType;
 import hu.unideb.smartcampus.shared.requestmessages.RetrieveSubjectsRequest;
